requirements.txt
django==4.2
social-auth-app-django
wikihowAPI-pk
django-webpack-loader
axios
djangorestframework
markdown
django-filter